+++ GameEngine/GameEngine/GameEngine.vcxproj
@@ -244,6 +244,8 @@
     <ClCompile Include="Engine\Components\DirectionalLightComponent.cpp" />
     <ClCompile Include="Engine\Components\Collider.cpp" />
     <ClCompile Include="Engine\Core\SceneChunkSerialization.cpp" />
+    <ClCompile Include="Engine\Entities\ArchetypeLibrary.cpp" />
+    <ClCompile Include="Engine\Entities\EntitySerializationInfos.cpp" />
     <ClCompile Include="Engine\Render\DepthMap.cpp" />
     <ClCompile Include="Engine\Render\DepthMapHandle.cpp" />
     <ClCompile Include="Engine\Render\DepthMapManager.cpp" />
@@ -432,6 +434,7 @@
     <ClInclude Include="Engine\Components\DirectionalLightComponent.hh" />
     <ClInclude Include="Engine\Components\Collider.hpp" />
     <ClInclude Include="Engine\Core\SceneChunkSerialization.hpp" />
+    <ClInclude Include="Engine\Entities\Archetype.hpp" />
     <ClInclude Include="Engine\Entities\ArchetypeLibrary.hpp" />
     <ClInclude Include="Engine\Render\DepthMap.hpp" />
     <ClInclude Include="Engine\Render\DepthMapHandle.hpp" />
