+++ GameEngine/Tools/AssetsEditor/AssetsEditor.vcxproj.filters
@@ -84,6 +84,12 @@
     <ClCompile Include="Convertor\MeshLoader.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Managers\ArchetypesEditorManager.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="EntityHelpers\EntityImgui.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Convertor\AssimpLoader.hpp">
